Fix package import after rename
