+++ pkg/state/rewards_test.go
@@ -1,4 +1,4 @@
-package rewards
+package state
 
 import (
 	"testing"
