+++ pkg/state_metrics/altair.go
@@ -1,10 +1,10 @@
-package reward_metrics
+package state_metrics
 
 import (
 	"math"
 
 	"github.com/attestantio/go-eth2-client/spec/phase0"
-	"github.com/cortze/eth2-state-analyzer/pkg/reward_metrics/fork_state"
+	"github.com/cortze/eth2-state-analyzer/pkg/state_metrics/fork_state"
 )
 
 type AltairMetrics struct {
