+++ pkg/state_metrics/bellatrix.go
@@ -1,1 +1,1 @@
-package reward_metrics
+package state_metrics
