+++ pkg/state_metrics/phase0.go
@@ -1,4 +1,4 @@
-package reward_metrics
+package state_metrics
 
 import (
 	"bytes"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/attestantio/go-eth2-client/spec/altair"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
-	"github.com/cortze/eth2-state-analyzer/pkg/reward_metrics/fork_state"
+	"github.com/cortze/eth2-state-analyzer/pkg/state_metrics/fork_state"
 )
 
 // var (
